auth_service/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator

